--- common/src/androidMain/kotlin/me/kevinschildhorn/common/viewmodel/EmailValidationViewModel.kt
@@ -1,5 +1,0 @@
-package me.kevinschildhorn.common.viewmodel
-
-import me.kevinschildhorn.android.viewModel.SharedEmailValidationViewModel
-
-class EmailValidationViewModel : SharedEmailValidationViewModel()
