+++ common/src/commonMain/kotlin/me/kevinschildhorn/common/layers/data/datasources/DirectoryDataSource.kt
@@ -1,0 +1,4 @@
+package me.kevinschildhorn.common.layers.data.datasources
+
+class DirectoryDataSource {
+}
