+++ common/src/commonMain/kotlin/me/kevinschildhorn/common/layers/data/repositories/CredentialsRepository.kt
@@ -1,0 +1,4 @@
+package me.kevinschildhorn.common.layers.data.repositories
+
+class CredentialsRepository {
+}
