+++ common/src/commonMain/kotlin/me/kevinschildhorn/common/layers/data/repositories/DirectoryRepository.kt
@@ -1,0 +1,7 @@
+package me.kevinschildhorn.common.layers.data.repositories
+
+import me.kevinschildhorn.common.layers.data.datasources.DirectoryDataSource
+
+class DirectoryRepository {
+    val dataSource: DirectoryDataSource = DirectoryDataSource()
+}
