+++ ios/FotoPresenter/swiftui/ContentView.swift
